added self-signed cert. added frontend single entry. added https support. fixed some bugs.

+++ core/internal/model/entity/account_role.go
@@ -2,8 +2,8 @@ package entity
 
 // AccountRole defines the account-role relationship entity
 type AccountRole struct {
-	Id         int `json:"id"          dc:"Relation ID"`
-	AccountId  int `json:"account_id"  dc:"Account ID"`
-	RoleId     int `json:"role_id"     dc:"Role ID"`
-	CreateTime int `json:"create_time" dc:"Creation time"`
+	Id         int64 `json:"id"          dc:"Relation ID"`
+	AccountId  int64 `json:"account_id"  dc:"Account ID"`
+	RoleId     int64 `json:"role_id"     dc:"Role ID"`
+	CreateTime int64 `json:"create_time" dc:"Creation time"`
 }
